+++ src/pdf.js
@@ -17,6 +17,7 @@
 import {
   addLinkAttributes,
   getFilenameFromUrl,
+  getPdfFilenameFromUrl,
   isFetchSupported,
   isPdfFile,
   isValidFetchUrl,
@@ -129,6 +130,7 @@ export {
   // From "./display/display_utils.js":
   addLinkAttributes,
   getFilenameFromUrl,
+  getPdfFilenameFromUrl,
   isPdfFile,
   LinkTarget,
   loadScript,
